Commit some intermediate shit matrix math that's broken
This is the part where you point and laugh, babe

+++ src/genesis/rendering/view/Camera.hpp
@@ -51,7 +51,7 @@ public:
         return orthoMatrix;
     }
 
-    glm::vec2 convertToWorld(int mouseX, int mouseY);
+    glm::vec2 convertToWorld(double mouseX, double mouseY);
 };
 
 } // namespace genesis
